backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pypdf
requests
openai
python-multipart
numpy
python-dotenv



